Atendimento dos requisitos constantes na issue refs #46
closes #46
