+++ tests/functional/seleniumPython/00-CriarSistema/test_00SEIRespostaCriarSistema.py
@@ -56,6 +56,28 @@ class Test00SEIRespostaCriarSistema():
     self.driver.find_element(By.ID, "txtServidor").send_keys("*")
     self.driver.find_element(By.ID, "txtServidor").send_keys(Keys.ENTER)
     self.driver.find_element(By.NAME, "sbmCadastrarServico").click()
+    self.driver.find_element(By.XPATH, "//img[@title=\'Operações\']").click()
+    self.driver.find_element(By.ID, "btnNovo").click()
+    dropdown = self.driver.find_element(By.ID, "selStaOperacaoServico")
+    dropdown.find_element(By.XPATH, "//option[. = 'Consultar Processo']").click()
+    self.driver.find_element(By.NAME, "sbmCadastrarOperacaoServico").click()
+    self.driver.find_element(By.ID, "btnNovo").click()
+    dropdown = self.driver.find_element(By.ID, "selStaOperacaoServico")
+    dropdown.find_element(By.XPATH, "//option[. = 'Enviar Processo']").click()
+    self.driver.find_element(By.NAME, "sbmCadastrarOperacaoServico").click()
+    self.driver.find_element(By.ID, "btnNovo").click()
+    dropdown = self.driver.find_element(By.ID, "selStaOperacaoServico")
+    dropdown.find_element(By.XPATH, "//option[. = 'Gerar Processo']").click()
+    self.driver.find_element(By.NAME, "sbmCadastrarOperacaoServico").click()
+    self.driver.find_element(By.ID, "btnNovo").click()
+    dropdown = self.driver.find_element(By.ID, "selStaOperacaoServico")
+    dropdown.find_element(By.XPATH, "//option[. = 'Incluir Documento']").click()
+    self.driver.find_element(By.NAME, "sbmCadastrarOperacaoServico").click()
+    self.driver.find_element(By.ID, "btnNovo").click()
+    dropdown = self.driver.find_element(By.ID, "selStaOperacaoServico")
+    dropdown.find_element(By.XPATH, "//option[. = 'Consultar Documento']").click()
+    self.driver.find_element(By.NAME, "sbmCadastrarOperacaoServico").click()
+    self.driver.find_element(By.ID, "btnFechar").click()
     self.driver.find_element(By.ID, "btnFechar").click()
   
   def test_criarTipoDocumento(self):
